requirements.txt
Flask
python-dotenv
ibm-watsonx-ai
requests